[PFM11-39] Use timer for the decay rate

+++ Source/ui/KnobWithLabels.cpp
@@ -69,5 +69,5 @@ void KnobWithLabels::paint (juce::Graphics& g)
 
 void KnobWithLabels::updateTooltip()
 {
-    setTooltip (std::to_string (getValue()));
+    setTooltip (std::to_string (static_cast<int>(getValue())));
 }
